Started checks and perfected space_utils
