+++ Inc/minishell.h
@@ -170,13 +170,15 @@ void	place_token(char *input, t_token **head);
 void	print_tokens(t_token *tokens);
 void	assign_type_token(t_token *token);
 void	assign_name(int type);
-void	check_types(t_token *token);
-void	checks_built_in(t_token *token);
-void	check_cmd_or_arg(t_token *token);
+void	assigns_types(t_token *token);
+void	assigns_built_in(t_token *token);
+void	assigns_cmd_or_arg(t_token *token);
 char	*add_spaces(char *input);
 int		space_length(char *input);
 char	*space_put(char *input, int len);
 void	check_quotes(char *input);
+void	clear_token_lst(t_token	**token);
+void	ft_error_check(t_token *token);
 // char	**cracked_split(char const *s, char c);
 // void	working_quote(char const *s, int *len, char c);
 // char	*extract_single_quote(const char *s, int len);
@@ -189,4 +191,4 @@ void	check_quotes(char *input);
 // FILE NAME IN ALL CAPS
 // functions in the file
 
-#endif
+#endif //MINISHELL_H
