+++ Src/Parse/tokenization.c
@@ -58,24 +58,6 @@ void	place_token(char *input, t_token **head)
 	}
 }
 
-/// @brief Assigns a Type to the Node
-/// @param token Node of a list 
-void	assign_type_token(t_token *token)
-{
-	t_token	*temp;
-
-	if(!token)
-		return ;
-	temp = token;
-	while (temp)
-	{
-		checks_built_in(temp);
-		check_types(temp);
-		check_cmd_or_arg(temp);
-		temp = temp->next;
-	}
-}
-
 /// @brief Adds spaces between operaters in the input
 /// @param input String received from the Stdout
 /// @return The new input with the spaces
@@ -87,6 +69,28 @@ char	*add_spaces(char *input)
 	len = space_length(input);
 	newinput = space_put(input, len);
 	printf("spaces: %d\n", len);
-	// printf("new_imput: %s\n", newinput);
 	return (newinput);
+}
+
+/// @brief Prints the node in order, there type and content
+/// @param tokens The token of the list 
+void	print_tokens(t_token *tokens)
+{
+	t_token	*curr;
+	int		i;
+
+	if (!tokens)
+		return ;
+	curr = tokens;
+	i = 0;
+	while (curr)
+	{
+		printf("Token[%d]: %s", i, curr->cont);
+		printf("\t");
+		assign_name(curr->type);
+		printf("\t");
+		printf("ID: %u\n", curr->type);
+		curr = curr->next;
+		i++;
+	}
 }
